+++ react-components/src/architecture/concrete/annotations/AnnotationsDomainObject.tsx
@@ -17,9 +17,9 @@ import { SingleAnnotation } from './helpers/SingleAnnotation';
 import { SolidDomainObject } from '../primitives/base/SolidDomainObject';
 import { PrimitiveType } from '../primitives/PrimitiveType';
 import { CylinderGizmoDomainObject } from './CylinderGizmoDomainObject';
-import { getStatusByAnnotation } from './utils/Status';
 import { AnnotationChangedDescription } from './helpers/AnnotationChangedDescription';
 import { DomainObjectChange } from '../../base/domainObjectsHelpers/DomainObjectChange';
+import { getStatusByAnnotation } from './helpers/Status';
 
 export class AnnotationsDomainObject extends VisualDomainObject {
   // ==================================================
