+++ react-components/src/architecture/concrete/annotations/utils/constants.ts
@@ -2,7 +2,4 @@
  * Copyright 2024 Cognite AS
  */
 
-export const FIT_CAMERA_TO_BOUNDING_BOX_ANIMATION_DURATION_MS = 1000;
-export const FIT_CAMERA_TO_BOUNDING_BOX_RADIUS_FACTOR = 5;
-
 export const ANNOTATION_CYLINDER_RADIUS_MARGIN = 0.2; // 20 % percent extra margin on the cylinder radius
